Refactor to make component more composable
