+++ ui/src/lib/queryBundle.utils.ts
@@ -1,5 +1,6 @@
 import { EntityRefKey } from '@/types';
 import { QueryEntities } from '~quent/types/QueryEntities';
+import { StatValue, RawNodeStatistics } from '@/services/query-plan/types';
 
 // Maps entity ref string to a key in the entities object.
 // Task has no corresponding collection in QueryEntities, so it is omitted.
@@ -21,4 +22,14 @@ export const ENTITY_REF_TO_ENTITIES_KEY: Partial<Record<EntityRefKey, keyof Quer
  */
 export function entityRefToEntitiesKey(entityRef: EntityRefKey): keyof QueryEntities | undefined {
   return ENTITY_REF_TO_ENTITIES_KEY[entityRef];
+}
+
+export function parseCustomStatistics(rawNode: unknown): Array<{ key: string; value: StatValue }> {
+  const statistics = (rawNode as RawNodeStatistics)?.statistics?.custom_statistics;
+  if (!statistics) return [];
+
+  return Object.entries(statistics).map(([key, tagged]) => ({
+    key,
+    value: Object.values(tagged)[0] ?? null,
+  }));
 }
